WIP: restructuring to make shapedetection and elementlocalization into one package "imageanalysis"
