+++ projectbambi/imageanalysis/__init__.py
@@ -1,0 +1,6 @@
+#imageanalysis
+
+def analyze_image(image): #'ipad3.jpg'
+    image = 'ipad3.jpg'
+    from .element_localization import xmlElements
+    return xmlElements
