+++ projectbambi/imageanalysis/shapedetector.py
